+++ chb/ast/ASTCPrettyPrinter.py
@@ -519,3 +519,12 @@ class ASTCPrettyPrinter(ASTVisitor):
 
     def visit_fieldinfo(self, finfo: AST.ASTFieldInfo) -> None:
         pass
+
+    def visit_enum_typ(self, t: AST.ASTTypEnum) -> None:
+        self.ccode.write("enum " + t.enumname)
+
+    def visit_enuminfo(self, einfo: AST.ASTEnumInfo) -> None:
+        self.ccode.write("enuminfo")
+
+    def visit_enumitem(self, eitem: AST.ASTEnumItem) -> None:
+        pass