+++ chb/ast/doc/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Version 0.1.0-2022-09-21
+
+- Added support for enum types: ASTEnumInfo, ASTEnumItem, ASTTypEnum
+- Added documentation to api for creating struct types
+
+---
 
 ### Version 0.1.0-2022-09-20